+++ MediTech/Dtos/UserDto.cs
@@ -8,9 +8,9 @@
         public string Telephone { get; set; }
         public string Address { get; set; }
         public string DateOfBirth { get; set; }
-        public int NationalID { get; set; }
+        public long NationalID { get; set; }
         public string Gender { get; set; }
-        public int? HealthInsuranceId { get; set; }
+        public long? HealthInsuranceId { get; set; }
         public string? HealthInsuranceCompany { get; set; }
         public string? Profile_Picture { get; set; }
         public string Token { get; set; }
